app/requirements.txt
streamlit
python-docx
kafka-python
openai
pyspark
scikit-learn
